Moving routing classes within a new subdirectory
