+++ tests/unit/src/Internal/Payment/Routing/RouterTest.php
@@ -12,8 +12,8 @@ use PHPUnit\Framework\MockObject\MockObject;
 use WCPay\Core\Server\Request\Get_Payment_Process_Factors;
 use WCPay\Database_Cache;
 use WCPay\Exceptions\API_Exception;
-use WCPay\Internal\Payment\Factor;
-use WCPay\Internal\Payment\Router;
+use WCPay\Internal\Payment\Routing\Factor;
+use WCPay\Internal\Payment\Routing\Router;
 
 /**
  * New payment process as a router router test.
